requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.8.0rc2
distlib==0.3.9
Django==5.1.4
django-allauth==65.3.1
django-cors-headers==4.7.0
django-debug-toolbar==4.4.6
django-environ==0.12.0
django-filter==24.3
django-jalali==7.4.0
django-jalali-date==2.0.0
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-nested-routers==0.94.1
factory_boy==3.3.1
Faker==33.1.0
filelock==3.16.1
idna==3.10
isodate==0.7.2
jalali_core==1.0.0
jdatetime==5.2.0
kavenegar==1.1.2
lxml==5.3.1
mysqlclient==2.2.6
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pipenv==2024.4.0
platformdirs==4.3.6
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.1
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
setuptools==75.8.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.28.1
wheel==0.45.1
zeep==4.3.1